feat: add temperature conversion
