--- data/constants.js
@@ -1,26 +1,0 @@
-export const familyData = [
-    {
-      role: "Père",
-      data: ["Stan", 45]
-    },
-    {
-      role: "Mère",
-      data: ["Francine", 45]
-    },
-    {
-      role: "Fille",
-      data: ["hayley", 18]
-    },
-    {
-      role: "Fils",
-      data: ["Steve", 14]
-    },
-    {
-      role: "Alien",
-      data: ["Roger", 1020]
-    },
-    {
-      role: "Poisson",
-      data: ["Klaus", 30]
-    },
-  ]
